backend/services/gemini_api.py
# services/gemini_api.py
from dotenv import load_dotenv
import os
import google.generativeai as genai

load_dotenv()

genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

def call_gemini(prompt):
    """
    Sends a prompt to the Gemini API and returns the generated SQL query.
    """
    model = genai.GenerativeModel('gemini-pro')
    response = model.generate_content(prompt)
    return response.text.strip()